+++ packages/netlify-cms-core/index.d.ts
@@ -308,6 +308,7 @@ declare module 'netlify-cms-core' {
     preview_path_date_field?: string;
     create?: boolean;
     delete?: boolean;
+    hide?: boolean;
     editor?: {
       preview?: boolean;
     };
@@ -482,7 +483,13 @@ declare module 'netlify-cms-core' {
 
   export interface CmsEventListener {
     name: 'prePublish' | 'postPublish' | 'preUnpublish' | 'postUnpublish' | 'preSave' | 'postSave';
-    handler: { entry: Map<string, any>; author: { login: string; name: string } };
+    handler: ({
+      entry,
+      author,
+    }: {
+      entry: Map<string, any>;
+      author: { login: string; name: string };
+    }) => any;
   }
 
   export type CmsEventListenerOptions = any; // TODO: type properly
@@ -510,9 +517,12 @@ declare module 'netlify-cms-core' {
     };
   }
 
-  type GetAssetFunction = (
-    asset: string,
-  ) => { url: string; path: string; field?: any; fileObj: File };
+  type GetAssetFunction = (asset: string) => {
+    url: string;
+    path: string;
+    field?: any;
+    fileObj: File;
+  };
 
   export type PreviewTemplateComponentProps = {
     entry: Map<string, any>;
@@ -525,6 +535,8 @@ declare module 'netlify-cms-core' {
     config: Map<string, any>;
     fields: List<Map<string, any>>;
     isLoadingAsset: boolean;
+    window: Window;
+    document: Document;
   };
 
   export interface CMS {
@@ -541,7 +553,7 @@ declare module 'netlify-cms-core' {
     registerEditorComponent: (options: EditorComponentOptions) => void;
     registerEventListener: (
       eventListener: CmsEventListener,
-      options: CmsEventListenerOptions,
+      options?: CmsEventListenerOptions,
     ) => void;
     registerLocale: (locale: string, phrases: CmsLocalePhrases) => void;
     registerMediaLibrary: (mediaLibrary: CmsMediaLibrary, options?: CmsMediaLibraryOptions) => void;