+++ packages/netlify-cms-core/src/__tests__/backend.spec.js
@@ -6,7 +6,7 @@ import {
   expandSearchEntries,
   mergeExpandedEntries,
 } from '../backend';
-import registry from '../lib/registry';
+import { getBackend } from '../lib/registry';
 import { FOLDER, FILES } from '../constants/collectionTypes';
 
 jest.mock('../lib/registry');
@@ -18,7 +18,7 @@ describe('Backend', () => {
     let backend;
 
     beforeEach(() => {
-      registry.getBackend.mockReturnValue({
+      getBackend.mockReturnValue({
         init: jest.fn(),
       });
       backend = resolveBackend({