+++ packages/netlify-cms-core/src/components/MediaLibrary/MediaLibraryModal.js
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import { Map } from 'immutable';
 import { isEmpty } from 'lodash';
 import { translate } from 'react-polyglot';
-import { Modal } from 'UI';
+import { Modal } from '../UI';
 import MediaLibraryTop from './MediaLibraryTop';
 import MediaLibraryCardGrid from './MediaLibraryCardGrid';
 import EmptyMessage from './EmptyMessage';