+++ packages/netlify-cms-core/src/lib/formatters.ts
@@ -182,7 +182,7 @@ export function previewUrlFormatter(
   let fields = entry.get('data') as Map<string, string>;
   fields = addFileTemplateFields(entry.get('path'), fields, collection.get('folder'));
   const dateFieldName = getDateField() || selectInferedField(collection, 'date');
-  const date = parseDateFromEntry((entry as unknown) as Map<string, unknown>, dateFieldName);
+  const date = parseDateFromEntry(entry as unknown as Map<string, unknown>, dateFieldName);
 
   // Prepare and sanitize slug variables only, leave the rest of the
   // `preview_path` template as is.
@@ -213,7 +213,7 @@ export function summaryFormatter(summaryTemplate: string, entry: EntryMap, colle
   let entryData = entry.get('data');
   const date =
     parseDateFromEntry(
-      (entry as unknown) as Map<string, unknown>,
+      entry as unknown as Map<string, unknown>,
       selectInferedField(collection, 'date'),
     ) || null;
   const identifier = entryData.getIn(keyToPathArray(selectIdentifier(collection) as string));
@@ -247,7 +247,7 @@ export function folderFormatter(
 
   const date =
     parseDateFromEntry(
-      (entry as unknown) as Map<string, unknown>,
+      entry as unknown as Map<string, unknown>,
       selectInferedField(collection, 'date'),
     ) || null;
   const identifier = fields.getIn(keyToPathArray(selectIdentifier(collection) as string));