+++ packages/netlify-cms-core/src/redux/index.ts
@@ -7,8 +7,8 @@ import { State } from '../types/redux';
 import { Reducer } from 'react';
 
 const store = createStore<State | undefined, AnyAction, unknown, unknown>(
-  (createRootReducer() as unknown) as Reducer<State | undefined, AnyAction>,
+  createRootReducer() as unknown as Reducer<State | undefined, AnyAction>,
   composeWithDevTools(applyMiddleware(thunkMiddleware as ThunkMiddleware<State>, waitUntilAction)),
 );
 
-export default store;
+export { store };