+++ packages/netlify-cms-lib-util/src/loadScript.js
@@ -7,7 +7,7 @@ export default function loadScript(url) {
     const head = document.getElementsByTagName('head')[0];
     const script = document.createElement('script');
     script.src = url;
-    script.onload = script.onreadystatechange = function() {
+    script.onload = script.onreadystatechange = function () {
       if (
         !done &&
         (!this.readyState || this.readyState === 'loaded' || this.readyState === 'complete')