+++ packages/netlify-cms-widget-markdown/src/MarkdownControl/plugins/ForceInsert.js
@@ -19,16 +19,10 @@ function ForceInsert({ defaultType }) {
       forceInsertBeforeNode(editor, node) {
         const block = { type: defaultType, object: 'block' };
         const parent = editor.value.document.getParent(node.key);
-        return editor
-          .insertNodeByKey(parent.key, 0, block)
-          .moveToStartOfNode(parent)
-          .focus();
+        return editor.insertNodeByKey(parent.key, 0, block).moveToStartOfNode(parent).focus();
       },
       forceInsertAfterNode(editor, node) {
-        return editor
-          .moveToEndOfNode(node)
-          .insertBlock(defaultType)
-          .focus();
+        return editor.moveToEndOfNode(node).insertBlock(defaultType).focus();
       },
       moveToEndOfDocument(editor) {
         const lastBlock = editor.value.document.nodes.last();