+++ packages/netlify-cms-widget-markdown/src/serializers/__tests__/index.spec.js
@@ -23,8 +23,7 @@ describe('markdownToSlate', () => {
             },
             {
               object: 'text',
-              text:
-                'this_mark, and your charge is but a penny; tothisa penny more; and so on to the full glass—the Cape Horn measure, which you may gulp down for a shilling.\\n\\nUpon entering the place I found a number of young seamen gathered about a table, examining by a dim light divers specimens ofskrimshander',
+              text: 'this_mark, and your charge is but a penny; tothisa penny more; and so on to the full glass—the Cape Horn measure, which you may gulp down for a shilling.\\n\\nUpon entering the place I found a number of young seamen gathered about a table, examining by a dim light divers specimens ofskrimshander',
               marks: [
                 {
                   type: 'italic',